Move configuration class from test and adjust tests
