+++ tests/base/test_configuration.py
@@ -2,11 +2,12 @@
 
 import unittest
 
-from tests.test_helper_config import (
-    ConfigurationOptionTestImpl,
-    ConfigurationTestImpl,
+from varats.base.configuration import (
+    DummyConfiguration,
+    ConfigurationImpl,
+    ConfigurationOptionImpl
 )
-from varats.base.configuration import DummyConfiguration
+from tests.test_utils import ConfigurationHelper
 
 
 class TestConfigurationOption(unittest.TestCase):
@@ -14,44 +15,44 @@ class TestConfigurationOption(unittest.TestCase):
 
     def test_basic_configuration_option_setup(self) -> None:
         """Test of the basic interface of configuration options works."""
-        config_option = ConfigurationOptionTestImpl("foo", 42)
+        config_option = ConfigurationOptionImpl("foo", 42)
 
         self.assertEqual(config_option.name, "foo")
         self.assertEqual(config_option.value, 42)
 
     def test_to_str(self) -> None:
         """Test to convert config option to string."""
-        config_option = ConfigurationOptionTestImpl("foo", 42)
+        config_option = ConfigurationOptionImpl("foo", 42)
 
         self.assertEqual(str(config_option), "foo: 42")
 
     def test_convert_to_bool(self) -> None:
         """Test to convert config option to bool."""
-        config_option_int = ConfigurationOptionTestImpl("foo", 42)
-        config_option_bool = ConfigurationOptionTestImpl("foo", False)
+        config_option_int = ConfigurationOptionImpl("foo", 42)
+        config_option_bool = ConfigurationOptionImpl("foo", False)
 
         self.assertEqual(bool(config_option_int), True)
         self.assertEqual(bool(config_option_bool), False)
 
     def test_equality_same(self) -> None:
         """Test to compare config option to to each other."""
-        config_option_int = ConfigurationOptionTestImpl("foo", 42)
-        config_option_int_2 = ConfigurationOptionTestImpl("foo", 42)
+        config_option_int = ConfigurationOptionImpl("foo", 42)
+        config_option_int_2 = ConfigurationOptionImpl("foo", 42)
 
         self.assertTrue(config_option_int == config_option_int_2)
         self.assertFalse(config_option_int != config_option_int_2)
 
     def test_equality_different(self) -> None:
         """Test to compare config option to to each other."""
-        config_option_int = ConfigurationOptionTestImpl("foo", 42)
-        config_option_bool = ConfigurationOptionTestImpl("foo", False)
+        config_option_int = ConfigurationOptionImpl("foo", 42)
+        config_option_bool = ConfigurationOptionImpl("foo", False)
 
         self.assertFalse(config_option_int == config_option_bool)
         self.assertTrue(config_option_int != config_option_bool)
 
     def test_equality_othertypes(self) -> None:
         """Test to compare config option to each other."""
-        config_option_int = ConfigurationOptionTestImpl("foo", 42)
+        config_option_int = ConfigurationOptionImpl("foo", 42)
         other_object = 42
 
         self.assertFalse(config_option_int == other_object)
@@ -63,7 +64,7 @@ class TestConfiguration(unittest.TestCase):
 
     def test_to_str(self) -> None:
         """Test to convert config to string."""
-        config = ConfigurationTestImpl.create_test_config()
+        config = ConfigurationHelper.create_test_config()
 
         self.assertEqual(str(config), config.dump_to_string())
 
@@ -78,7 +79,7 @@ class TestDummyConfiguration(unittest.TestCase):
     def test_crash_add_config_option(self) -> None:
         with self.assertRaises(AssertionError):
             d_config = DummyConfiguration()
-            d_config.add_config_option(ConfigurationOptionTestImpl("foo", 42))
+            d_config.add_config_option(ConfigurationOptionImpl("foo", 42))
 
     def test_crash_set_config_option(self) -> None:
         with self.assertRaises(AssertionError):
