+++ tests/mapping/test_configuration_map.py
@@ -5,11 +5,11 @@ import unittest.mock as mock
 from pathlib import Path
 from tempfile import NamedTemporaryFile
 
-from tests.test_helper_config import (
-    ConfigurationTestImpl,
-    ConfigurationOptionTestImpl,
+from varats.base.configuration import (
+    DummyConfiguration,
+    ConfigurationImpl,
+    ConfigurationOptionImpl
 )
-from varats.base.configuration import DummyConfiguration
 from varats.mapping.configuration_map import (
     ConfigurationMap,
     store_configuration_map,
@@ -23,7 +23,7 @@ Version: 1
 ...
 """
 YAML_DOC_CONFIG_MAP = """---
-0: '{''foo'': ''foo: True'', ''bar'': ''bar: False'', ''bazz'': ''bazz: bazz-value''}'
+0: '{'foo': 'True', 'bar': 'False', 'bazz': 'bazz-value'}'
 1: '{}'
 2: '{}'
 ...
@@ -36,7 +36,7 @@ class TestConfigurationMap(unittest.TestCase):
     def test_add_get_config(self) -> None:
         """Tests if we can add and retrieve a config from the map."""
         config_map = ConfigurationMap()
-        test_config = ConfigurationTestImpl()
+        test_config = ConfigurationImpl()
 
         config_id = config_map.add_configuration(test_config)
 
@@ -57,9 +57,9 @@ class TestConfigurationMap(unittest.TestCase):
     def test_add_get_multiple_configs(self) -> None:
         """Tests if we can add and retrieve multiple configs from the map."""
         config_map = ConfigurationMap()
-        test_config_1 = ConfigurationTestImpl()
-        test_config_2 = ConfigurationTestImpl()
-        test_config_3 = ConfigurationTestImpl()
+        test_config_1 = ConfigurationImpl()
+        test_config_2 = ConfigurationImpl()
+        test_config_3 = ConfigurationImpl()
 
         config_id_1 = config_map.add_configuration(test_config_1)
         config_id_2 = config_map.add_configuration(test_config_2)
@@ -82,9 +82,9 @@ class TestConfigurationMap(unittest.TestCase):
     def test_inter_configs(self) -> None:
         """Test if we can iterate over all configurations."""
         config_map = ConfigurationMap()
-        test_config_1 = ConfigurationTestImpl()
-        test_config_2 = ConfigurationTestImpl()
-        test_config_3 = ConfigurationTestImpl()
+        test_config_1 = ConfigurationImpl()
+        test_config_2 = ConfigurationImpl()
+        test_config_3 = ConfigurationImpl()
 
         config_map.add_configuration(test_config_1)
         config_map.add_configuration(test_config_2)
@@ -97,9 +97,9 @@ class TestConfigurationMap(unittest.TestCase):
     def test_inter_id_config_tuples(self) -> None:
         """Test if we can iterate over all id configuration pairs."""
         config_map = ConfigurationMap()
-        test_config_1 = ConfigurationTestImpl()
-        test_config_2 = ConfigurationTestImpl()
-        test_config_3 = ConfigurationTestImpl()
+        test_config_1 = ConfigurationImpl()
+        test_config_2 = ConfigurationImpl()
+        test_config_3 = ConfigurationImpl()
 
         config_map.add_configuration(test_config_1)
         config_map.add_configuration(test_config_2)
@@ -118,9 +118,9 @@ class TestConfigurationMapStoreAndLoad(unittest.TestCase):
     def setUpClass(cls):
         """Setup test ConfigurationMap."""
         cls.config_map = ConfigurationMap()
-        cls.test_config_1 = ConfigurationTestImpl.create_test_config()
-        cls.test_config_2 = ConfigurationTestImpl()
-        cls.test_config_3 = ConfigurationTestImpl()
+        cls.test_config_1 = ConfigurationImpl.create_test_config()
+        cls.test_config_2 = ConfigurationImpl()
+        cls.test_config_3 = ConfigurationImpl()
 
         cls.config_id_1 = cls.config_map.add_configuration(cls.test_config_1)
         cls.config_id_2 = cls.config_map.add_configuration(cls.test_config_2)
@@ -149,7 +149,7 @@ class TestConfigurationMapStoreAndLoad(unittest.TestCase):
         ]
 
         config_map = load_configuration_map(
-            Path("fake_file_path"), ConfigurationTestImpl
+            Path("fake_file_path"), ConfigurationImpl
         )
 
         self.assertSetEqual({0, 1, 2}, set(config_map.ids()))
@@ -160,17 +160,17 @@ class TestConfigurationMapStoreAndLoad(unittest.TestCase):
         yaml doc."""
         config_map = create_configuration_map_from_yaml_doc({
             '0':
-                "{'foo': 'foo: True', 'bar': 'bar: False', "
-                "'bazz': 'bazz: bazz-value'}",
+                "{'foo': 'True', 'bar': 'False', "
+                "'bazz': 'bazz-value'}",
             '1': "{}"
-        }, ConfigurationTestImpl)
+        }, ConfigurationImpl)
 
         self.assertSetEqual({0, 1}, set(config_map.ids()))
         config = config_map.get_configuration(0)
         self.assertTrue(config is not None)
         if config:
             self.assertEqual(
-                ConfigurationOptionTestImpl("foo", True),
+                ConfigurationOptionImpl("foo", True),
                 config.get_config_value("foo")
             )
         self.assertTrue(config_map.get_configuration(1) is not None)
